Add private array containing 2 float values
This is necessary to make Vector2f useable with OpenGL methods that take
in a float pointer, especially methods with 2fv appended to the end of
their name (e.g. glVertex2fv, glTexCoord2fv), because if individual
values are used instead of a vector, you cannot turn them into an array
and use it.
Even if you were to turn them into an array and return a pointer to it,
the array will be gone after the method exits because the array is a
local variable.