Change to getter instead of const reference
Initially, we decided to go with a const reference so as to make it look
very closely to Unity's Time.deltaTime, which doesn't need the
invocation operator ().

While the syntax is more familiar, it posed a problem for the other
Time variables. Suppose we want to have a way to obtain the time passed
since the beginning of the program, we can't do it conveniently using
the const ref method.
If we did it using our const ref method,we have to
ask for the time since we started the program at every frame, update a
private variable (say "m_time"), then hook it up via a reference.

While essentially this lets us use Time.time, there are a few drawbacks:
1. Time.time now doesn't exactly tell us the time since the program
   started, but the time since the program started before this frame.
   If you get Time.time, perform some intensive task, then get
   Time.time again and compare both values, they'd be the same, though
   they should be different.
2. It is unnecessary to demand for Time.time again and again when you
   don't use it. It'd be better to just get the time on demand.