Change constructor to take in const references
If we use the previous way, the following occurs:
1. Vectors passed in via that constructor is copied into the
   constructor.
2. When initializing the members, the copy is then copied again and
   given to the member fields.

By passing in by value, we wasted one round of copying because the
vectors will be copied again when we initialize the member fields.

To prevent this, we can make the parameters constant references.
References do not copy the vector because they refer to a specific
vector. We make the references constant to prevent code from changing
its value and hence, affect the variable passed in (it is a reference,
so any changes made to the reference will affect the original variable).