Add unordered map to store each key's status
Added an unordered map to store each key's status. The KeyStatus struct
is mapped to an integer, which will come from the enum class KeyCode.

You can obtain the integer value by casting from the enum element to an
int.

I also used an unordered_map because order is not needed here. By not
requiring the need for the map to be sorted, inserting elements to it is
much faster (average constant time complexity -> O(1)).