Add methods to propagate the key down/up status
Added methods to send key down/up events to the unordered map in the
Input class.

When a key is pressed down/up, the value of wParam is compared against
the unordered map. To do that, we simply use the at() method. This
retrieves the value at the specified key which is the value of wParam.

Once the value is retrieved, we modify its members (isHeldDown, isDown,
isUp).