Handle out_of_range exception in send key methods
Previously, when sending a value of wParam that doesn't exist in the
unordered_map (mainly because we forgot to include key codes for the
rest of the keys), the program will crash with an std::out_of_range
exception.

Now, the std::out_of_range exception is caught, which just makes the
method return. This may be potentially dangerous because this is making
the program throw away an exception silently. Should make it output
something I think?