Add a constructor for the Material class
Added a constructor for the Material class that takes in 4 arguments,
a Color4f for the ambient, a Color4f for the diffuse, a Color4f for the
specular and a GLfloat for the shininess.

The constructor assigns its member arrays using the values in the
Color4f. We require an array of GLfloats because
the glMaterialfv requires it.