Add another private vector to the Scene class
The vector will contain all GameObjects that have been added to the
scene during Start and Update. Because adding to the standard
"gameObjects" vector will invalidate iterators, we have to do it in
another vector.

At the end of the Update loop, all elements in the newly-added
GameObjects vector will be transferred to the standard GameObjects
vector. The newly-added GameObjects vector will then be cleared for
future use.