Update AddComponent using program phases
If the program is in the initialization phase, the components will be
directly added to the vector of components.

If the program is running however, the components must go through a
different vector, which will be responsible for calling the Start method
on each of the components in the vector.