Add unoptimized of picking pickable meshes
Kudos to Tiong Guan for fixing the picking for me! Turns out I forgot
to accommodate the width and height of the title bar of the window!

Added a method called DrawPickableMeshes in the Program class that draws
all pickable meshes to the back buffer, reads the back buffer to an
array then checks to see if the color of the pixel at the mouse's
position is equal to that of the color of the mesh when drawn to the
back buffer.