Use member variable to store back buffer image
Added a member variable in Program that can store the back buffer
image. This is a much needed improvement over declaring this as a local
variable in a method, where we will be allocating and deallocating it as
we click, which may impact performance.

I originally don't know the syntax for declaring a multidimensional
array that will be dynamically allocated until I read this
Stack Overflow comment:
https://stackoverflow.com/questions/936687/how-do-i-declare-a-2d-array-in-c-using-new#comment744797_936687